# apps/api/openapi.json
{
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "LoadOut": {
        "properties": {
          "offerings": {
            "title": "Offerings",
            "type": "integer"
          }
        },
        "required": [
          "offerings"
        ],
        "title": "LoadOut",
        "type": "object"
      },
      "ManualPriceIn": {
        "properties": {
          "cacheReadPerMtok": {
            "anyOf": [
              {
                "minimum": 0.0,
                "type": "number"
              },
              {
                "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Cachereadpermtok"
          },
          "cacheWritePerMtok": {
            "anyOf": [
              {
                "minimum": 0.0,
                "type": "number"
              },
              {
                "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Cachewritepermtok"
          },
          "inputPerMtok": {
            "anyOf": [
              {
                "minimum": 0.0,
                "type": "number"
              },
              {
                "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$",
                "type": "string"
              }
            ],
            "title": "Inputpermtok"
          },
          "outputPerMtok": {
            "anyOf": [
              {
                "minimum": 0.0,
                "type": "number"
              },
              {
                "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$",
                "type": "string"
              }
            ],
            "title": "Outputpermtok"
          },
          "requestUsd": {
            "anyOf": [
              {
                "minimum": 0.0,
                "type": "number"
              },
              {
                "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Requestusd"
          }
        },
        "required": [
          "inputPerMtok",
          "outputPerMtok"
        ],
        "title": "ManualPriceIn",
        "type": "object"
      },
...
      "OfferingOut": {
        "properties": {
          "allowedByPolicy": {
            "title": "Allowedbypolicy",
            "type": "boolean"
          },
          "capabilities": {
            "items": {
              "type": "string"
            },
            "title": "Capabilities",
            "type": "array"
          },
          "contextWindow": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Contextwindow"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "maxOutputTokens": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Maxoutputtokens"
          },
          "policyReason": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Policyreason"
          },
          "price": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/PriceOut"
              },
              {
                "type": "null"
              }
            ]
          },
          "status": {
            "enum": [
              "available",
              "unavailable"
            ],
            "title": "Status",
            "type": "string"
          },
          "upstreamProvider": {
            "title": "Upstreamprovider",
            "type": "string"
          },
          "zdr": {
            "title": "Zdr",
            "type": "boolean"
          }
        },
        "required": [
          "id",
          "upstreamProvider",
          "contextWindow",
          "maxOutputTokens",
          "capabilities",
          "zdr",
          "status",
          "price",
          "allowedByPolicy",
          "policyReason"
        ],
        "title": "OfferingOut",
        "type": "object"
      },
...
      "PriceOut": {
        "properties": {
          "cacheReadPerMtok": {
            "anyOf": [
              {
                "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Cachereadpermtok"
          },
          "cacheWritePerMtok": {
            "anyOf": [
              {
                "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Cachewritepermtok"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "inputPerMtok": {
            "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$",
            "title": "Inputpermtok",
            "type": "string"
          },
          "outputPerMtok": {
            "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$",
            "title": "Outputpermtok",
            "type": "string"
          },
          "requestUsd": {
            "anyOf": [
              {
                "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Requestusd"
          },
          "source": {
            "enum": [
              "provider_sync",
              "manual"
            ],
            "title": "Source",
            "type": "string"
          },
          "validFrom": {
            "format": "date-time",
            "title": "Validfrom",
            "type": "string"
          },
          "validTo": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Validto"
          }
        },
        "required": [
          "id",
          "inputPerMtok",
          "outputPerMtok",
          "cacheReadPerMtok",
          "cacheWritePerMtok",
          "requestUsd",
          "validFrom",
          "validTo",
          "source"
        ],
        "title": "PriceOut",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
      "VersionOut": {
        "properties": {
          "canonicalSlug": {
            "title": "Canonicalslug",
            "type": "string"
          },
          "capabilities": {
            "items": {
              "type": "string"
            },
            "title": "Capabilities",
            "type": "array"
          },
          "contextWindow": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Contextwindow"
          },
          "family": {
            "title": "Family",
            "type": "string"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "offerings": {
            "items": {
              "$ref": "#/components/schemas/OfferingOut"
            },
            "title": "Offerings",
            "type": "array"
          },
          "providerSlug": {
            "title": "Providerslug",
            "type": "string"
          },
          "status": {
            "enum": [
              "available",
              "deprecated"
            ],
            "title": "Status",
            "type": "string"
          }
        },
        "required": [
          "id",
          "family",
          "providerSlug",
          "canonicalSlug",
          "name",
          "contextWindow",
          "capabilities",
          "status",
          "offerings"
        ],
        "title": "VersionOut",
        "type": "object"
      },
...
    }
  },
  "info": {
    "title": "NexTI Platform API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/ai/catalog": {
      "get": {
        "operationId": "list_catalog_api_v1_ai_catalog_get",
        "parameters": [
          {
            "in": "query",
            "name": "search",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Search"
            }
          },
          {
            "in": "query",
            "name": "only_offered",
            "required": false,
            "schema": {
              "default": false,
              "title": "Only Offered",
              "type": "boolean"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/VersionOut"
                  },
                  "title": "Response List Catalog Api V1 Ai Catalog Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Catalog",
        "tags": [
          "ai"
        ]
      }
    },
    "/api/v1/ai/catalog/versions/{version_id}:load-offerings": {
      "post": {
        "description": "Load the upstream providers of one model with their current prices.",
        "operationId": "load_api_v1_ai_catalog_versions__version_id__load_offerings_post",
        "parameters": [
          {
            "in": "path",
            "name": "version_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Version Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LoadOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Load",
        "tags": [
          "ai"
        ]
      }
    },
...
    "/api/v1/ai/offerings/{offering_id}/prices": {
      "get": {
        "operationId": "list_prices_api_v1_ai_offerings__offering_id__prices_get",
        "parameters": [
          {
            "in": "path",
            "name": "offering_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Offering Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/PriceOut"
                  },
                  "title": "Response List Prices Api V1 Ai Offerings  Offering Id  Prices Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Prices",
        "tags": [
          "ai"
        ]
      },
      "post": {
        "description": "A price the provider does not report (spec 12.2). Past costs keep the version they used.",
        "operationId": "add_manual_price_api_v1_ai_offerings__offering_id__prices_post",
        "parameters": [
          {
            "in": "path",
            "name": "offering_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Offering Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ManualPriceIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/PriceOut"
                  },
                  "title": "Response Add Manual Price Api V1 Ai Offerings  Offering Id  Prices Post",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Add Manual Price",
        "tags": [
          "ai"
        ]
      }
    },
...
  }
}